chatbot/requirements.txt
langchain==0.0.202
aim==4.0.0a13
click
python-dotenv
openai
google-search-results
